Renamed Go scanner package to "scan" for consistency.
